Split up the mappers
