+++ lib/mappers/index.js
@@ -1,0 +1,7 @@
+const mapTextStyle = require('./textstyle');
+const mapColor = require('./color');
+
+module.exports = {
+  mapColors: colors => colors.map(mapColor),
+  mapTextStyles: textStyles => textStyles.objects.map(mapTextStyle),
+};
